+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added Field Output Handler for outputting a text when a contact has a certain relationship.
 * Refactored Field Output Handlers and simplified duplicate code for initializing a field.
 * Added Is not empty as an operator to the contact filter.
+* Added field to edit activity.
 
 # Version 1.10.0